Refactor SharedConstantReflectionProvider so it becomes a superclass for both hosted and runtime implementations.
