+++ substratevm/src/com.oracle.svm.core.graal.amd64/src/com/oracle/svm/core/graal/amd64/AMD64ImageHeapAddressOptimizationPhase.java
@@ -33,14 +33,12 @@ import java.util.ListIterator;
 import org.graalvm.nativeimage.Platform;
 import org.graalvm.nativeimage.Platforms;
 
-import com.oracle.graal.pointsto.heap.ImageHeapRelocatableConstant;
 import com.oracle.svm.core.Isolates;
 import com.oracle.svm.core.ReservedRegisters;
 import com.oracle.svm.core.SubstrateOptions;
-import com.oracle.svm.shared.util.SubstrateUtil;
-import com.oracle.svm.shared.feature.AutomaticallyRegisteredFeature;
 import com.oracle.svm.core.feature.InternalFeature;
 import com.oracle.svm.core.graal.meta.SharedConstantReflectionProvider;
+import com.oracle.svm.shared.feature.AutomaticallyRegisteredFeature;
 import com.oracle.svm.shared.option.HostedOptionKey;
 import com.oracle.svm.shared.util.VMError;
 
@@ -124,19 +122,7 @@ class AMD64ImageHeapAddressOptimizationPhase extends BasePhase<CoreProviders> {
     }
 
     static boolean canOptimize(CompressibleConstant constant, ConstantReflectionProvider constantReflection) {
-        if (!phaseEnabled()) {
-            return false;
-
-        } else if (SubstrateUtil.HOSTED) {
-            /*
-             * ImageHeapRelocatableConstants are placeholders for values patched during execution
-             * startup and hence are not compatible with this optimization.
-             */
-            return !(constant instanceof ImageHeapRelocatableConstant);
-
-        } else {
-            return ((SharedConstantReflectionProvider) constantReflection).getImageHeapOffset(constant) != 0;
-        }
+        return phaseEnabled() && ((SharedConstantReflectionProvider) constantReflection).canRepresentAsImageHeapOffset(constant);
     }
 
     @Override
@@ -159,7 +145,7 @@ class AMD64ImageHeapAddressOptimizationPhase extends BasePhase<CoreProviders> {
                     continue;
                 }
 
-                if (AMD64ImageHeapAddressOptimizationPhase.canOptimize(baseConstant, constantReflection)) {
+                if (canOptimize(baseConstant, constantReflection)) {
                     AMD64ImageHeapAddressNode replacement = new AMD64ImageHeapAddressNode(baseConstant, address.getIndex(), address.getScale(), address.getDisplacement());
                     address.replaceAndDelete(graph.unique(replacement));
                     if (baseNode.hasNoUsages()) {
