+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/graal/meta/SharedConstantReflectionProvider.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2023, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2026, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -24,85 +24,20 @@
  */
 package com.oracle.svm.core.graal.meta;
 
-import static com.oracle.svm.shared.util.VMError.shouldNotReachHereAtRuntime;
-
-import java.lang.reflect.Array;
-import java.util.function.ObjIntConsumer;
-
-import com.oracle.svm.core.meta.ObjectConstantEquality;
-import com.oracle.svm.core.meta.SubstrateObjectConstant;
 import com.oracle.svm.shared.util.VMError;
 
 import jdk.vm.ci.meta.Constant;
 import jdk.vm.ci.meta.ConstantReflectionProvider;
 import jdk.vm.ci.meta.JavaConstant;
-import jdk.vm.ci.meta.JavaKind;
-import jdk.vm.ci.meta.MethodHandleAccessProvider;
 import jdk.vm.ci.meta.ResolvedJavaType;
 
 public abstract class SharedConstantReflectionProvider implements ConstantReflectionProvider {
 
-    @Override
-    public Boolean constantEquals(Constant x, Constant y) {
-        if (x == y) {
-            return true;
-        } else if (x instanceof SubstrateObjectConstant) {
-            return y instanceof SubstrateObjectConstant && ObjectConstantEquality.get().test((SubstrateObjectConstant) x, (SubstrateObjectConstant) y);
-        } else {
-            return x.equals(y);
-        }
-    }
-
-    @Override
-    public Integer readArrayLength(JavaConstant array) {
-        if (array.getJavaKind() != JavaKind.Object || array.isNull() || !SubstrateObjectConstant.asObject(array).getClass().isArray()) {
-            return null;
-        }
-        return java.lang.reflect.Array.getLength(SubstrateObjectConstant.asObject(array));
-    }
-
-    @Override
-    public JavaConstant readArrayElement(JavaConstant array, int index) {
-        if (array.getJavaKind() != JavaKind.Object || array.isNull()) {
-            return null;
-        }
-
-        Object a = SubstrateObjectConstant.asObject(array);
-
-        if (!a.getClass().isArray() || index < 0 || index >= Array.getLength(a)) {
-            return null;
-        }
-
-        if (a instanceof Object[]) {
-            Object element = ((Object[]) a)[index];
-            return forObject(element);
-        } else {
-            return JavaConstant.forBoxedPrimitive(Array.get(a, index));
-        }
-    }
-
-    public void forEachArrayElement(JavaConstant array, ObjIntConsumer<JavaConstant> consumer) {
-        if (array.getJavaKind() != JavaKind.Object || array.isNull()) {
-            return;
-        }
-
-        Object obj = SubstrateObjectConstant.asObject(array);
-
-        if (!obj.getClass().isArray()) {
-            return;
-        }
-
-        if (obj instanceof Object[] a) {
-            for (int index = 0; index < a.length; index++) {
-                consumer.accept((forObject(a[index])), index);
-            }
-        } else {
-            for (int index = 0; index < Array.getLength(obj); index++) {
-                Object element = Array.get(obj, index);
-                consumer.accept(JavaConstant.forBoxedPrimitive((element)), index);
-            }
-        }
-    }
+    /**
+     * Determines whether {@code constant} refers to an image heap object with an offset that is
+     * known or can be patched once known.
+     */
+    public abstract boolean canRepresentAsImageHeapOffset(JavaConstant constant);
 
     @Override
     public final JavaConstant boxPrimitive(JavaConstant source) {
@@ -112,28 +47,6 @@ public abstract class SharedConstantReflectionProvider implements ConstantReflec
          * give you an Integer box instead of a Byte, Short, ... box.
          */
         throw VMError.intentionallyUnimplemented();
-    }
-
-    @Override
-    public JavaConstant unboxPrimitive(JavaConstant source) {
-        if (!source.getJavaKind().isObject()) {
-            return null;
-        }
-        return JavaConstant.forBoxedPrimitive(SubstrateObjectConstant.asObject(source));
-    }
-
-    @Override
-    public JavaConstant forString(String value) {
-        return SubstrateObjectConstant.forObject(value);
-    }
-
-    protected JavaConstant forObject(Object object) {
-        return SubstrateObjectConstant.forObject(object);
-    }
-
-    @Override
-    public MethodHandleAccessProvider getMethodHandleAccess() {
-        throw shouldNotReachHereAtRuntime(); // ExcludeFromJacocoGeneratedReport
     }
 
     @Override
