+++ substratevm/src/com.oracle.svm.hosted.test/src/com/oracle/svm/hosted/test/VerifyReflectionUsage.java
@@ -121,7 +121,6 @@ public class VerifyReflectionUsage extends VerifyReflectionUsageBase {
                     clazz("com.oracle.svm.core.graal.llvm.runtime.LLVMExceptionUnwind"),
                     clazz("com.oracle.svm.core.graal.meta.DynamicHubOffsets"),
                     clazz("com.oracle.svm.core.graal.meta.KnownOffsets"),
-                    clazz("com.oracle.svm.core.graal.meta.SharedConstantReflectionProvider"),
                     clazz("com.oracle.svm.core.graal.meta.SubstrateReplacements"),
                     clazz("com.oracle.svm.core.graal.meta.SubstrateReplacements$SnippetInlineInvokePlugin"),
                     clazz("com.oracle.svm.core.graal.meta.SubstrateSnippetReflectionProvider"),
@@ -376,6 +375,7 @@ public class VerifyReflectionUsage extends VerifyReflectionUsageBase {
                     clazz("com.oracle.svm.core.jdk.StackWalkerUtil", SVM_RUN_TIME_ONLY),
                     clazz("com.oracle.svm.core.jdk.Target_java_net_URL$DefaultFactory", SVM_RUN_TIME_ONLY),
                     clazz("com.oracle.svm.core.jfr.JfrTypeRepository", SVM_RUN_TIME_ONLY),
+                    clazz("com.oracle.svm.graal.meta.SubstrateConstantReflectionProvider", SVM_RUN_TIME_ONLY),
                     pkg("com.oracle.svm.guest", GUEST_CONTEXT_ONLY),
                     method("com.oracle.svm.hosted.imagelayer.HostedImageLayerBuildingSupport", "validateSingletonRegistration", NI_HOSTED_IMPLEMENTATION),
                     clazz("com.oracle.svm.hosted.ModuleLayerFeature", TERMINUS_OBSOLETED),
