+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/ameta/AnalysisConstantReflectionProvider.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, 2017, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2026, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -45,6 +45,7 @@ import com.oracle.graal.pointsto.meta.AnalysisField;
 import com.oracle.graal.pointsto.meta.AnalysisType;
 import com.oracle.graal.pointsto.meta.AnalysisUniverse;
 import com.oracle.svm.core.classinitialization.TypeReachedProvider;
+import com.oracle.svm.core.graal.meta.SharedConstantReflectionProvider;
 import com.oracle.svm.core.hub.DynamicHub;
 import com.oracle.svm.hosted.SVMHost;
 import com.oracle.svm.hosted.classinitialization.ClassInitializationSupport;
@@ -55,7 +56,6 @@ import com.oracle.svm.util.GuestAccess;
 
 import jdk.vm.ci.code.CodeUtil;
 import jdk.vm.ci.meta.Constant;
-import jdk.vm.ci.meta.ConstantReflectionProvider;
 import jdk.vm.ci.meta.JavaConstant;
 import jdk.vm.ci.meta.JavaKind;
 import jdk.vm.ci.meta.MemoryAccessProvider;
@@ -65,7 +65,7 @@ import jdk.vm.ci.meta.ResolvedJavaField;
 import jdk.vm.ci.meta.ResolvedJavaType;
 
 @Platforms(Platform.HOSTED_ONLY.class)
-public class AnalysisConstantReflectionProvider implements ConstantReflectionProvider, TypeReachedProvider {
+public class AnalysisConstantReflectionProvider extends SharedConstantReflectionProvider implements TypeReachedProvider {
     private final AnalysisUniverse universe;
     protected final UniverseMetaAccess metaAccess;
     private final AnalysisMethodHandleAccessProvider methodHandleAccess;
@@ -78,6 +78,20 @@ public class AnalysisConstantReflectionProvider implements ConstantReflectionPro
         this.metaAccess = metaAccess;
         this.methodHandleAccess = new AnalysisMethodHandleAccessProvider(universe);
         this.classInitializationSupport = classInitializationSupport;
+    }
+
+    @Override
+    public boolean canRepresentAsImageHeapOffset(JavaConstant constant) {
+        /*
+         * Relocatable constants are placeholders patched during execution startup rather than
+         * ordinary image heap constants whose offsets can be patched after image heap layout.
+         */
+        return constant instanceof ImageHeapConstant && !(constant instanceof ImageHeapRelocatableConstant);
+    }
+
+    @Override
+    public int getImageHeapOffset(JavaConstant constant) {
+        throw VMError.shouldNotReachHere("Image heap offsets are only available during JIT compilation at run time: " + getClass().getName());
     }
 
     @Override
@@ -273,11 +287,6 @@ public class AnalysisConstantReflectionProvider implements ConstantReflectionPro
         return readValue((AnalysisField) field, receiver, false, false);
     }
 
-    @Override
-    public JavaConstant boxPrimitive(JavaConstant source) {
-        throw VMError.intentionallyUnimplemented();
-    }
-
     public JavaConstant readValue(AnalysisField field, JavaConstant receiver, boolean returnSimulatedValues, boolean readRelocatableValues) {
         if (!field.isStatic()) {
             if (!(receiver instanceof ImageHeapInstance imageHeapInstance) || !field.getDeclaringClass().isAssignableFrom(imageHeapInstance.getType())) {
@@ -384,15 +393,6 @@ public class AnalysisConstantReflectionProvider implements ConstantReflectionPro
     @Override
     public JavaConstant asJavaClass(ResolvedJavaType type) {
         return universe.getHeapScanner().createImageHeapConstant(getHostVM().dynamicHub(type), ObjectScanner.OtherReason.UNKNOWN);
-    }
-
-    @Override
-    public Constant asObjectHub(ResolvedJavaType type) {
-        /*
-         * Substrate VM does not distinguish between the hub and the Class, they are both
-         * represented by the DynamicHub.
-         */
-        return asJavaClass(type);
     }
 
     @Override
